+++ Tc3Build/Tc3Build/Program.cs
@@ -1,6 +1,5 @@
 using System.CommandLine;
 using Tc3Build.Commands;
-using Tc3Build.Infrastructure;
 
 namespace Tc3Build;
 
@@ -9,15 +8,7 @@ internal static class Program
     [STAThread]
     private static int Main(string[] args)
     {
-        ComMessageFilter.Register();
-        try
-        {
-            var rootCommand = RootCommandFactory.Create();
-            return rootCommand.Parse(args).Invoke();
-        }
-        finally
-        {
-            ComMessageFilter.Revoke();
-        }
+        var rootCommand = RootCommandFactory.Create();
+        return rootCommand.Parse(args).Invoke();
     }
 }